requirements.txt
a2wsgi==1.10.8
aiobotocore==2.19.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
asyncssh==2.19.0
atpublic==5.0
attrs==24.3.0
babel==2.17.0
bentoml==1.4.0
billiard==4.2.1
blinker==1.9.0
boto3==1.36.3
botocore==1.36.3
cachelib==0.9.0
cachetools==5.5.2
cattrs==23.1.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
configobj==5.0.9
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.44.1
deepmerge==2.0
Deprecated==1.2.18
deprecation==2.1.0
dictdiffer==0.9.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.0
dvc-data==3.16.8
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
dynaconf==3.2.10
email_validator==2.2.0
entrypoints==0.4
evidently==0.6.5
Faker==36.1.1
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.56.0
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.12.0
funcy==2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-re2==1.1.20240702
googleapis-common-protos==1.68.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.70.0
gto==1.7.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-ws==0.7.1
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
inflection==0.5.1
iterative-telemetry==0.0.9
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.6
kantoku==0.18.1
kiwisolver==1.4.8
kombu==5.4.2
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
litestar==2.14.0
litestar-htmx==0.4.1
lockfile==0.12.2
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.20.2
mlflow-skinny==2.20.2
more-itertools==10.6.0
msgspec==0.19.0
multidict==6.1.0
multipart==1.2.1
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
nvidia-ml-py==12.570.86
omegaconf==2.3.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-aiohttp-client==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
patsy==1.0.1
pendulum==3.0.0
pillow==11.1.0
pip-requirements-parser==32.0.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polyfactory==2.19.0
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
questionary==2.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rich-click==1.8.6
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2024.12.0
s3transfer==0.11.2
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
scmrepo==3.3.9
semver==3.0.2
setproctitle==1.3.4
setuptools==75.8.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simple-di==0.1.5
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
sqltrie==0.11.1
starlette==0.45.3
statsmodels==0.14.4
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tomli_w==1.2.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
ujson==5.10.0
universal_pathlib==0.2.6
urllib3==2.3.0
uuid6==2024.7.10
uv==0.6.2
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
